+++ src/js/helpers/event-bus.js
@@ -80,7 +80,7 @@ export class EventBus extends Vue {
       createAssetRequestsUpdateList: 'createAssetRequests:updateList',
       atomicSwapsUpdateList: 'atomicSwaps:updateList',
       updateAccountRole: 'updateAccountRole',
-      promocodesUpdateList: 'promocodes:updateList',
+      promoCodesUpdateList: 'promoCodes:updateList',
     }
   }