+++ src/main.js
@@ -19,7 +19,7 @@ import { formatMoney } from '@/vue/filters/formatMoney'
 import { formatBalance } from './vue/filters/formatBalance'
 import { formatNumber } from '@/vue/filters/formatNumber'
 import { formatInteger } from '@/vue/filters/formatInteger'
-import { formatToPercent } from '@/vue/filters/formatToPercent'
+import { formatPercent } from '@/vue/filters/formatPercent'
 import { formatCalendar } from '@/vue/filters/formatCalendar'
 import { formatCalendarInline } from '@/vue/filters/formatCalendarInline'
 import { formatDateDMY } from '@/vue/filters/formatDateDMY'
@@ -45,7 +45,7 @@ async function init () {
   Vue.filter('formatMoney', formatMoney)
   Vue.filter('formatBalance', formatBalance)
   Vue.filter('formatNumber', formatNumber)
-  Vue.filter('formatToPercent', formatToPercent)
+  Vue.filter('formatPercent', formatPercent)
   Vue.filter('formatInteger', formatInteger)
   Vue.filter('formatCalendar', formatCalendar)
   Vue.filter('formatCalendarInline', formatCalendarInline)