+++ src/vue/filters/formatPercent.js
@@ -1,5 +1,5 @@
 import { globalize } from './globalize'
 
-export function formatToPercent (value) {
+export function formatPercent (value) {
   return globalize('formats.percent', { value })
 }