+++ src/vue/forms/PayForm.vue
@@ -63,6 +63,7 @@ export default {
         amount: '',
         quoteAssetCode: '',
         paymentMethodId: '',
+        promoCode: '',
       },
       isDisabled: false,
       atomicSwapBidDetails: {
@@ -110,7 +111,7 @@ export default {
           this.form.amount,
           this.form.paymentMethodId,
           this.atomicSwapAsk.id,
-          this.form.promocode,
+          this.form.promoCode,
           this.form.email
         )
         if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {