+++ src/vue/modules/atomic-swap-form/index.vue
@@ -62,6 +62,7 @@ export default {
         amount: '',
         quoteAssetCode: '',
         paymentMethodId: '',
+        promoCode: '',
       },
       isDisabled: false,
       atomicSwapBidDetails: {
@@ -92,7 +93,7 @@ export default {
           this.form.amount,
           this.form.paymentMethodId,
           this.atomicSwapAsk.id,
-          this.form.promocode
+          this.form.promoCode
         )
         if (atomicSwapBid.type === ATOMIC_SWAP_BID_TYPES.redirect) {
           window.location.href = atomicSwapBid.payUrl