+++ src/vue/navigation/Sidebar.vue
@@ -149,14 +149,14 @@
             v-ripple
             class="sidebar__link"
             @click.native="closeSidebar"
-            :to="vueRoutes.promocodes"
+            :to="vueRoutes.promoCodes"
             tag="a"
           >
             <i
               class="sidebar__link-icon mdi mdi-ticket-percent"
             />
             <span>
-              {{ 'pages-names.promocodes' | globalize }}
+              {{ 'pages-names.promo-codes' | globalize }}
             </span>
           </router-link>
         </nav>